refactor: AIFrameworkDashboard を適切なモジュール構造に分割
693行の単一ファイルを以下の構造に整理:
- layers/: レイヤーコンポーネント群 (6ファイル)
- tabs/: タブコンテンツ群 (4ファイル)
- ui/: 共通UIコンポーネント
- types.ts: 型定義
- index.tsx: メインコンポーネント

メインファイル(AIFrameworkDashboard.tsx)は11行のre-exportに
削減量: 682行 (目標393行を超過達成)
