+++ components/AIFrameworkDashboard/tabs/index.ts
@@ -1,0 +1,4 @@
+export { PromptTab } from './PromptTab';
+export { FlowTab } from './FlowTab';
+export { ParamsTab } from './ParamsTab';
+export { DebugTab } from './DebugTab';
